moved some things into inner classes to clean up a bit
