+++ src/main/java/voltskiya/apple/game_mechanics/tmw/tmw_config/biomes/gui/BiomeTypeGuiPageMobs.java
@@ -7,14 +7,10 @@ import voltskiya.apple.game_mechanics.tmw.tmw_config.TMWGui;
 import voltskiya.apple.game_mechanics.tmw.tmw_config.biomes.BiomeType;
 import voltskiya.apple.game_mechanics.tmw.tmw_config.mobs.MobType;
 import voltskiya.apple.game_mechanics.tmw.tmw_config.mobs.gui.MobTypeGui;
-import voltskiya.apple.game_mechanics.tmw.tmw_config.mobs.gui.MobTypeInventorySlot;
-import voltskiya.apple.game_mechanics.util.gui.InventoryGui;
 import voltskiya.apple.game_mechanics.util.gui.InventoryGuiPageScrollable;
 import voltskiya.apple.game_mechanics.util.gui.InventoryGuiSlotGeneric;
 import voltskiya.apple.game_mechanics.util.gui.InventoryGuiSlotScrollable;
 import voltskiya.apple.game_mechanics.util.minecraft.InventoryUtils;
-
-import java.util.Collections;
 
 public class BiomeTypeGuiPageMobs extends InventoryGuiPageScrollable {
     private final BiomeTypeGui biomeTypeGui;
