+++ src/main/java/voltskiya/apple/game_mechanics/tmw/tmw_config/mobs/MobType.java
@@ -1,16 +1,21 @@
 package voltskiya.apple.game_mechanics.tmw.tmw_config.mobs;
 
 import com.google.gson.*;
+import net.minecraft.server.v1_16_R3.EntityTypes;
+import net.minecraft.server.v1_16_R3.NBTTagCompound;
+import org.bukkit.Material;
 import org.bukkit.inventory.ItemStack;
-import voltskiya.apple.game_mechanics.tmw.tmw_config.mobs.gui.MobTypeBuilder;
+import org.bukkit.inventory.meta.ItemMeta;
+import voltskiya.apple.game_mechanics.util.minecraft.InventoryUtils;
 
 import java.lang.reflect.Type;
+import java.util.List;
 
 public class MobType {
     private final MobTypeBuilder.MobIcon icon;
 
-    public MobType(MobTypeBuilder.MobIcon icon) {
-        this.icon = icon;
+    public MobType(MobTypeBuilder builder) {
+        this.icon = builder.icon;
     }
 
     public ItemStack toItem() {
@@ -23,10 +28,6 @@ public class MobType {
 
     public String getName() {
         return icon.getName();
-    }
-
-    public MobTypeBuilder.MobIcon getIcon() {
-        return icon;
     }
 
     @Override
@@ -44,16 +45,68 @@ public class MobType {
         return this.icon.getName();
     }
 
-    public static class MobTypeSerializer implements JsonSerializer<MobType>{
+    public static class MobTypeSerializer implements JsonSerializer<MobType> {
         @Override
         public JsonElement serialize(MobType mobType, Type type, JsonSerializationContext jsonSerializationContext) {
             return new JsonPrimitive(mobType.icon.getName());
         }
     }
-    public static class MobTypeDeSerializer implements  JsonDeserializer<MobType> {
+
+    public static class MobTypeDeSerializer implements JsonDeserializer<MobType> {
         @Override
         public MobType deserialize(JsonElement jsonElement, Type type, JsonDeserializationContext jsonDeserializationContext) throws JsonParseException {
             return MobTypeDatabase.getMob(jsonElement.getAsString());
+        }
+    }
+
+    public static class MobTypeBuilder {
+        private MobIcon icon;
+
+        public MobTypeBuilder(MobType real) {
+            this.icon = real.icon;
+        }
+
+        public MobTypeBuilder() {
+            icon = null;
+        }
+
+        public void setIcon(MobIcon icon) {
+            this.icon = icon;
+        }
+
+        public ItemStack getIconItem() {
+            return icon == null ? InventoryUtils.makeItem(Material.LEVER, 1, "No spawn egg", null) : icon.toItem();
+        }
+
+        public MobType build() {
+            return new MobType(this);
+        }
+
+        public static class MobIcon {
+            private final String name;
+            private final Material material;
+            private final List<String> lore;
+            private final NBTTagCompound nbt;
+
+            public MobIcon(String name, Material material, List<String> lore, NBTTagCompound nbt, EntityTypes<?> entityTypes) {
+                this.name = name;
+                this.material = material;
+                this.lore = lore;
+                this.nbt = nbt;
+            }
+
+            public ItemStack toItem() {
+                ItemStack item = new ItemStack(material);
+                final ItemMeta itemMeta = item.getItemMeta();
+                itemMeta.setDisplayName(name);
+                itemMeta.setLore(lore);
+                item.setItemMeta(itemMeta);
+                return item;
+            }
+
+            public String getName() {
+                return name;
+            }
         }
     }
 }
