+++ src/main/java/voltskiya/apple/game_mechanics/tmw/tmw_config/mobs/gui/MobTypeGui.java
@@ -1,10 +1,11 @@
 package voltskiya.apple.game_mechanics.tmw.tmw_config.mobs.gui;
 
 import voltskiya.apple.game_mechanics.tmw.tmw_config.TMWGui;
+import voltskiya.apple.game_mechanics.tmw.tmw_config.mobs.MobType;
 import voltskiya.apple.game_mechanics.util.gui.InventoryGui;
 
 public class MobTypeGui extends InventoryGui {
-    public MobTypeGui(TMWGui tmwGui,MobTypeBuilder mob) {
+    public MobTypeGui(TMWGui tmwGui, MobType.MobTypeBuilder mob) {
         addPage(
                 new MobTypeGuiPageSettings(this,mob,tmwGui),
                 new MobTypeGuiPageBiomes(this,mob),
