+++ src/main/java/voltskiya/apple/game_mechanics/tmw/tmw_config/mobs/gui/MobTypeGuiPageBiomes.java
@@ -1,16 +1,16 @@
 package voltskiya.apple.game_mechanics.tmw.tmw_config.mobs.gui;
 
 import org.bukkit.Material;
-import voltskiya.apple.game_mechanics.util.gui.InventoryGui;
+import voltskiya.apple.game_mechanics.tmw.tmw_config.mobs.MobType;
 import voltskiya.apple.game_mechanics.util.gui.InventoryGuiPageScrollable;
 import voltskiya.apple.game_mechanics.util.gui.InventoryGuiSlotGeneric;
 import voltskiya.apple.game_mechanics.util.minecraft.InventoryUtils;
 
 public class MobTypeGuiPageBiomes extends InventoryGuiPageScrollable {
     private MobTypeGui mobTypeGui;
-    private MobTypeBuilder mob;
+    private MobType.MobTypeBuilder mob;
 
-    public MobTypeGuiPageBiomes(MobTypeGui mobTypeGui, MobTypeBuilder mob) {
+    public MobTypeGuiPageBiomes(MobTypeGui mobTypeGui, MobType.MobTypeBuilder mob) {
         super(mobTypeGui);
         this.mobTypeGui = mobTypeGui;
         this.mob = mob;
